Mark the type of a line by a prefix of the line instead of by sections.
Every line has a prefix string to identify the type of the data of the
line.  Previous, the data type of lines are identified by sections.
For example, in "MAPS" or "FLOWS" section.

The new way is more easy to generate maps entries incrementally,
instead of write out all maps entries at beginning one shot.